add ability to specify which job to run

+++ sync-node/src/main/java/io/trxplorer/syncnode/job/NodeSyncJob.java
@@ -7,6 +7,7 @@ import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.mashape.unirest.http.exceptions.UnirestException;
 
+import io.trxplorer.syncnode.SyncNodeConfig;
 import io.trxplorer.syncnode.service.NodeSyncService;
 
 @Singleton
@@ -14,14 +15,20 @@ import io.trxplorer.syncnode.service.NodeSyncService;
 public class NodeSyncJob {
 	
 	private NodeSyncService nodeSyncService;
+	private SyncNodeConfig config;
 
 	@Inject
-	public NodeSyncJob(NodeSyncService nodeSyncService) {
+	public NodeSyncJob(NodeSyncService nodeSyncService,SyncNodeConfig config) {
 		this.nodeSyncService = nodeSyncService;
+		this.config = config;
 	}
 	
 	@Scheduled("1m")
 	public void addNewNodes() throws InterruptedException{
+		
+		if (!this.config.isNodesJobEnabled()) {
+			return;
+		}
 		
 		this.nodeSyncService.syncNodes();
 		
@@ -31,18 +38,27 @@ public class NodeSyncJob {
 	@Scheduled("1m")
 	public void updateNodesStatus() {
 		
+		if (!this.config.isNodesJobEnabled()) {
+			return;
+		}
+		
 		this.nodeSyncService.updateNodesStatus();
 		
 	}
 
 	@Scheduled("1m")
 	public void removeDownNodes() {
+		if (!this.config.isNodesJobEnabled()) {
+			return;
+		}
 		this.nodeSyncService.removeDownNodes();
 	}
 	
 	@Scheduled("5m")
 	public void updateNodeLocalization() throws UnirestException {
-		
+		if (!this.config.isNodesJobEnabled()) {
+			return;
+		}
 		this.nodeSyncService.updateNodesLocalizationData();
 		
 	}
